+++ tests/Application.IntegrationTests/Testing.cs
@@ -1,5 +1,5 @@
 ﻿using CleanArchitecture.Infrastructure.Identity;
-using CleanArchitecture.Infrastructure.Persistence;
+using CleanArchitecture.Infrastructure.Data;
 using MediatR;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc.Testing;
